+++ server/controllers/todoController.js
@@ -41,7 +41,8 @@ module.exports = class TodoController {
             title: req.body.title,
             description: req.body.description,
             status: req.body.status,
-            due_date: req.body.due_date
+            due_date: req.body.due_date,
+            UserId: req.user
         }
         Todo.create(newData)
         .then( data => {
