+++ server/models/user.js
@@ -12,7 +12,9 @@ module.exports = (sequelize, DataTypes) => {
      * The `models/index` file will call this method automatically.
      */
     static associate(models) {
-      // define association here
+      User.hasMany(models.Todo, {
+        foreignKey: 'UserId'
+      })
     }
   };
   User.init({
